+++ pages/scheduledFormulas.php
@@ -376,3 +376,4 @@ setInterval(monitorAndReloadScheduledTable, 5000);
 
 
 <script src="/js/import.making.js"></script>
+<script src="/js/pvMakingApp.js"></script>