+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Version 13.2
 - Formula making api added "Complete Formula"
 - Added Scheduled formula auto table update
+- Formula making Android app link update
+- Formula making app detection improvements
 
 ### Version 13.1
 - Dashboard pie color is retained in local storage